requirements39.txt
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
ibm-cloud-sdk-core==3.20.1
ibm-cos-sdk==2.11.0
ibm-cos-sdk-core==2.11.0
ibm-cos-sdk-s3transfer==2.11.0
ibm-db==3.2.0
ibm-watson==8.1.0
ibm-watson-machine-learning==1.0.333
ibm-watsonx-ai==0.0.5
openai==1.31.0